adds folders, stylesheets for components moved out of Channel, renames Controls to TrackControls
