+++ src/components/editor/controls/Dropdown/Dropdown.css
@@ -1,0 +1,10 @@
+.addChannelSelect {
+  border-radius: 10px;
+  border: var(--white-ish) double 5px;
+  font-family: 'Barlow';
+  align-self: center;
+  width: 180px;
+  height: 60px;
+  background-color: var(--black-ish);
+  color: var(--white-ish);
+}
