+++ src/components/editor/controls/TrackControls/TrackControls.css
@@ -1,0 +1,18 @@
+.controls {
+  font-size: 1rem;
+  text-align: center;
+  max-width: 8rem;
+}
+
+.controls label {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  max-width: 100%;
+}
+
+.controls input[type='range'] {
+  display: block;
+  width: 100%;
+  max-width: fit-content;
+}
