+++ src/components/editor/controls/index.js
@@ -1,8 +1,8 @@
 // handles exports of all components in current folder
 
 import Channel from './Channel/Channel.jsx';
-import Controls from './Channel/Controls.jsx';
-import Dropdown from './Channel/Dropdown.jsx';
+import Controls from './TrackControls/TrackControls.jsx';
+import Dropdown from './Dropdown/Dropdown.jsx';
 import GlobalControls from './GlobalControls/GlobalControls.jsx';
 
 export { Channel, Controls, Dropdown, GlobalControls };
